+++ contracts/btc-finality/src/contract.rs
@@ -16,6 +16,7 @@ use crate::state::finality::get_btc_staking_activated_height;
 use crate::state::finality::ACCUMULATED_VOTING_WEIGHTS;
 use crate::{finality, queries, state};
 use babylon_apis::finality_api::SudoMsg;
+use cosmwasm_logging::{debug, info, init_cosmwasm_logger};
 use cosmwasm_std::{
     attr, to_json_binary, Addr, Deps, DepsMut, Env, MessageInfo, QueryResponse, Reply, Response,
     StdResult,
@@ -33,6 +34,11 @@ pub fn instantiate(
     info: MessageInfo,
     msg: InstantiateMsg,
 ) -> Result<Response, ContractError> {
+    init_cosmwasm_logger(deps.api);
+
+    info!("Instantiating BTC finality contract");
+    debug!("Instantiate message: {:?}", msg);
+
     nonpayable(&info)?;
     let denom = deps.querier.query_bonded_denom()?;
 
@@ -69,6 +75,8 @@ pub fn reply(_deps: DepsMut, _env: Env, _reply: Reply) -> StdResult<Response> {
 }
 
 pub fn query(deps: Deps, _env: Env, msg: QueryMsg) -> Result<QueryResponse, ContractError> {
+    init_cosmwasm_logger(deps.api);
+
     match msg {
         QueryMsg::Config {} => Ok(to_json_binary(&CONFIG.load(deps.storage)?)?),
         QueryMsg::Admin {} => to_json_binary(&ADMIN.query_admin(deps)?).map_err(Into::into),
@@ -139,6 +147,8 @@ pub fn query(deps: Deps, _env: Env, msg: QueryMsg) -> Result<QueryResponse, Cont
 /// This function is called when the contract is migrated to a new version.
 /// For non-state-breaking migrations, this updates the contract version and logs the migration.
 pub fn migrate(deps: DepsMut, _env: Env, _msg: MigrateMsg) -> Result<Response, ContractError> {
+    init_cosmwasm_logger(deps.api);
+
     // Get the current version stored in the contract
     let prev_version = cw2::get_contract_version(deps.storage)?;
 
@@ -165,6 +175,7 @@ pub fn execute(
     info: MessageInfo,
     msg: ExecuteMsg,
 ) -> Result<Response, ContractError> {
+    init_cosmwasm_logger(deps.api);
     let api = deps.api;
     match msg {
         ExecuteMsg::UpdateAdmin { admin } => ADMIN